requirements.txt
anthropic
feedparser
pyyaml
